requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.3.0
click==8.1.3
Django==4.2.2
djangorestframework==3.14.0
drf-spectacular==0.26.3
flake8==6.0.0
inflection==0.5.1
jsonschema==4.17.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==10.0.0
platformdirs==3.8.0
pycodestyle==2.10.0
pyflakes==3.0.1
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
uritemplate==4.1.1
